+++ src/Voice.js
@@ -1,36 +1,50 @@
 "use strict";
 
+import nexmo from './index';
+
 class Voice {
   /**
    * @param {Credentials} credentials
    *    credentials to be used when interacting with the API.
    * @param {Object} options
-   *    Addition Voice options.
+   *    Addition  options.
    */
   constructor(credentials, options = {}) {
     this.creds = credentials;
     this.options = options;
+    
+    // Used to facilitate testing of the call to the underlying object
+    this._nexmo = this.options.nexmoOverride || nexmo;
+    
+    this._nexmo.initialize(this.creds.apiKey, this.creds.apiSecret, this.options);
   }
   
   /**
-   * Make a call using the Nexmo Voice API.
-   * For full details of parameters see https://docs.nexmo.com/voice/voice-api/api-reference#call_create
-   * @param {Object} params - Call parameters
-   * @param {function} callback - callback executed when either an error occurs
-   *                      or the request to the API succeeds.
+   * TODO: document
    */
-  call(params, callback) {
-    var config = {
-      host:'api.nexmo.com',
-      path:'/beta/calls',
-      method: 'POST',
-      body: JSON.stringify(params),
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${this.creds.generateJwt()}`
-      }
-    };
-    this.options.httpClient.request(config, callback);
+  sendTTSMessage() {
+    this._nexmo.sendTTSMessage.apply(this._nexmo, arguments);
+  }
+  
+  /**
+   * TODO: document
+   */
+  sendTTSPromptWithCapture() {
+    this._nexmo.sendTTSPromptWithCapture.apply(this._nexmo, arguments);
+  }
+  
+  /**
+   * TODO: document
+   */
+  sendTTSPromptWithConfirm() {
+    this._nexmo.sendTTSPromptWithConfirm.apply(this._nexmo, arguments);
+  }
+  
+  /**
+   * TODO: document
+   */
+  call() {
+    this._nexmo.call.apply(this._nexmo, arguments);
   }
     
 }
